remove func sig functions to util.functions
