--- bokeh/util/future.py
@@ -1,107 +1,0 @@
-#-----------------------------------------------------------------------------
-# Copyright (c) 2012 - 2019, Anaconda, Inc., and Bokeh Contributors.
-# All rights reserved.
-#
-# The full license is in the file LICENSE.txt, distributed with this software.
-#-----------------------------------------------------------------------------
-''' Utilities for Py2/Py3 interop.
-
-'''
-
-#-----------------------------------------------------------------------------
-# Boilerplate
-#-----------------------------------------------------------------------------
-from __future__ import absolute_import, division, print_function
-
-import logging
-log = logging.getLogger(__name__)
-
-#-----------------------------------------------------------------------------
-# Imports
-#-----------------------------------------------------------------------------
-
-# Standard library imports
-import sys
-
-# External imports
-
-# Bokeh imports
-
-#-----------------------------------------------------------------------------
-# Globals and constants
-#-----------------------------------------------------------------------------
-
-__all__ = (
-    'format_signature',
-    'get_param_info',
-)
-
-#-----------------------------------------------------------------------------
-# General API
-#-----------------------------------------------------------------------------
-
-#-----------------------------------------------------------------------------
-# Dev API
-#-----------------------------------------------------------------------------
-
-#-----------------------------------------------------------------------------
-# Private API
-#-----------------------------------------------------------------------------
-
-#-----------------------------------------------------------------------------
-# Code
-#-----------------------------------------------------------------------------
-
-# inspect.getargspec and inspect.formatargspec were deprecated in Python 3.5
-# in favor of the newer inspect.signature introspection
-
-if sys.version_info[:2] < (3, 4):
-
-    def signature(func):
-        # The modifications in this function are to make results more in line
-        # with Python 3, i.e. self is not included in bound methods, supplied
-        # parameters are not reported in partial, etc. This simplifies the
-        # downstream code considerably.
-        from inspect import getargspec, isfunction, ismethod
-        from functools import partial
-
-        if isfunction(func) or ismethod(func):
-            sig = getargspec(func)
-            if ismethod(func):
-                sig.args.remove('self')
-            return sig
-
-        elif isinstance(func, partial):
-            sig = getargspec(func.func)
-            if 'self' in sig.args: sig.args.remove('self')
-            if func.keywords is not None:
-                for name in func.keywords.keys():
-                    sig.args.remove(name)
-            for val in func.args:
-                del sig.args[0]
-            return sig
-
-        else:
-            sig = getargspec(func.__call__)
-            sig.args.remove('self')
-            return sig
-
-    def format_signature(sig):
-        from inspect import formatargspec
-        return formatargspec(*sig)
-
-    def get_param_info(sig):
-        return (sig.args, sig.defaults or [])
-
-else:
-    from inspect import signature; signature
-
-    def format_signature(sig):
-        return str(sig)
-
-    def get_param_info(sig):
-        defaults = []
-        for param in sig.parameters.values():
-            if param.default is not param.empty:
-                defaults.append(param.default)
-        return list(sig.parameters), defaults
